create a dedicated user 'ng-users' in the master realm
This user is to be used by the shanoir-ng-users microservice when
updating the keycloak database (add/remove/update a user). Its
permissions are limited to managing the users of the shanoir-ng
realm.

The purpose of this change is to avoid giving total admin access
to the keycloak server from shanoir-ng-users (a bit too critical)
and decrease the coupling between the two containers (the config
of shanoir-ng-keycloak shall be managed by itself, not by another
container).

-> can manage users of the shanoir-ng realm