allow prepending a prefix to the container names
All container names are prefixes with the value of the environment
variable SHANOIR_PREFIX (if set). This is to allow running multiple
instances of shanoir on the same docker host.

NOTE: for the nginx container, this requires to generate the config
nginx.conf dynamically at runtime (the new generate config now honours
the SHANOIR_URL_SCHEME and SHANOIR_URL_HOST vars).