shanoir-ng-post-auth keycloak authenticator
This commit merges all keycloak additions into a single keycloak
authenticator named `shanoir-ng-post-auth` and adds the ability to
restrict admin logins from a whitelist of IP addresses or networks.

The authenticator is designed to be run at the end of the authentication
flows, in the `REQUIRED` mode.

It performs 3 actions:
- ensure that the account has not expired (this replaces the
  `CheckExpirationDate` script which relied on the deprecated Nashorn
  engine)
- ensure that users with `ROLE_ADMIN` log from an IP listed in
  `SHANOIR_ALLOWED_ADMIN_IPS`
- send a request to the users ms to update the last login date (this
  replaces the `RequiredActionProvider`)

It is appended to the `browser`, `direct grant`, `http challenge` and
`reset credential` authentication flows.

Notes:
- it renames the flows (eg: `browser` -> `shanoir-ng browser`) and reset
  their `builtIn` flag so that keycloak does not tamper with them on
  import.
- it adds a new dependency on the `com.github.seancfoley:ipaddress`
  package (to handle IP addresses)