fix the detection of existing migrations
The "check migration names" stage in the ci pipeline fails to
detect existing migrations in the master branch. They appear as
`(none)` in https://github.com/fli-iam/shanoir-ng/actions/runs/3046116136/jobs/4908518031

With proper detection this script would have caught the bugs
that had to be fixed in ca84e9502b and d25400f693.