Fix for repository == null with Mockito only,
The issue: repository in AcqEquipServiceImpl + another repo in BasicEntityServiceImpl., that was never replaced by Mockito as hierarchy only (so, no way to move to SpringBootTest for a simple service test, so I keep Mockito only tests, as light and right level), but as no solution for empty repository with calls to e.g. findById, that did all not work, I moved down the methods of BasicEntityService, what makes a duplication in some way, but at least the tests work and the schema is more simple. I kept the interface hierarchy.